Consolidated kubernetes funcs into single package.
